requirements.txt
opencv-python==4.8.1.78
PyPDF2==3.0.1
Pillow==10.0.1
numpy==1.24.3
mediapipe==0.10.7
pygame==2.5.2
PyMuPDF==1.23.5